+++ packages/typespec-ts/src/modular/helpers/typeHelpers.ts
@@ -101,7 +101,7 @@ export function getType(type: Type, format?: string): TypeMetadata {
     case "dict":
       return handleDictType(type);
 
-    case "never": 
+    case "never":
       return { name: "never", nullable: type.nullable };
 
     default: